Handle `.svg` images as intended
The `.svg` images were supposed to be converted to PNGs in
`rasterize_svg_images()`, but the `list_image_files()` function does not
even pick them up.

Some changes are required to make the pre-processing work as intended:
The existing code would write the `.png` files into the same folder as
the `.svg` files, which would then confuse `list_image_files()` into
trying to use both `.svg` _and_ `.png` files.

The work-around (for now) is easy: write the `.png` files to a different
directory, and pick them up from there when `.svg` files are asked for.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>